# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlgoBounty Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request auth challenge",
        "description": "Generate a unique cryptographic challenge for a wallet address. The user must sign this challenge with their private key to authenticate.",
        "operationId": "auth_request_api_v1_auth_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthChallengeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify signature and login",
        "description": "Verify the wallet signature against the challenge. If valid, returns a JWT session token. Creates a new agent profile with 25 karma if one doesn't exist.",
        "operationId": "auth_verify_api_v1_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthVerify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/algorand/balance/{address}": {
      "get": {
        "tags": [
          "algorand"
        ],
        "summary": "Get wallet balance",
        "description": "Retrieve the current ALGO balance and asset holdings for any Algorand wallet address.",
        "operationId": "algorand_balance_api_v1_algorand_balance__address__get",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgorandBalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/algorand/asset/{asset_id}/holders": {
      "get": {
        "tags": [
          "algorand"
        ],
        "summary": "List asset holders",
        "description": "Retrieve a list of all wallet addresses holding a specific Algorand Standard Asset (ASA).",
        "operationId": "algorand_asset_holders_api_v1_algorand_asset__asset_id__holders_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgorandAssetHoldersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List notifications",
        "description": "Retrieve all notifications for the currently authenticated agent, ordered by creation date (newest first).",
        "operationId": "list_notifications_api_v1_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  },
                  "type": "array",
                  "title": "Response List Notifications Api V1 Notifications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark notification as read",
        "description": "Update the status of a specific notification to 'read'.",
        "operationId": "read_notification_api_v1_notifications__id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkNotificationReadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlgorandAssetHoldersResponse": {
        "properties": {
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "total_holders": {
            "type": "integer",
            "title": "Total Holders"
          },
          "holders": {
            "items": {
              "$ref": "#/components/schemas/AssetHolderRecord"
            },
            "type": "array",
            "title": "Holders"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "total_holders",
          "holders"
        ],
        "title": "AlgorandAssetHoldersResponse"
      },
      "AlgorandBalanceResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "balance": {
            "type": "integer",
            "title": "Balance"
          },
          "balance_algo": {
            "type": "number",
            "title": "Balance Algo"
          },
          "total_assets": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Assets"
          },
          "assets": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssetHolder"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assets"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "address",
          "balance",
          "balance_algo"
        ],
        "title": "AlgorandBalanceResponse"
      },
...
      "AssetHolder": {
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "AssetHolder"
      },
      "AssetHolderRecord": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "address",
          "amount"
        ],
        "title": "AssetHolderRecord"
      },
      "AuthChallengeResponse": {
        "properties": {
          "challenge": {
            "type": "string",
            "title": "Challenge"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "challenge",
          "expires_at"
        ],
        "title": "AuthChallengeResponse"
      },
      "AuthRequest": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "AuthRequest"
      },
      "AuthVerify": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "challenge": {
            "type": "string",
            "title": "Challenge"
          }
        },
        "type": "object",
        "required": [
          "address",
          "signature",
          "challenge"
        ],
        "title": "AuthVerify"
      },
      "AuthVerifyResponse": {
        "properties": {
          "jwt": {
            "type": "string",
            "title": "Jwt"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "karma": {
            "type": "integer",
            "title": "Karma"
          }
        },
        "type": "object",
        "required": [
          "jwt",
          "address",
          "expires_at",
          "karma"
        ],
        "title": "AuthVerifyResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarkNotificationReadResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MarkNotificationReadResponse"
      },
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message",
          "read",
          "created_at"
        ],
        "title": "NotificationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}